# neuralblitz-v50/go/pkg/api/server.go
package api

import (
	"net/http"
	"runtime"
	"time"

	"github.com/gin-gonic/gin"
	"neuralblitz/pkg/core"
	"neuralblitz/pkg/options"
	"neuralblitz/pkg/utils"
)

// Server represents the API server
type Server struct {
	router      *gin.Engine
	dyad        *core.ArchitectSystemDyad
	engine      *core.SelfActualizationEngine
	interpreter *options.NBCLInterpreter
	port        string
	startTime   time.Time
}

// NewServer creates a new API server
func NewServer(port string) *Server {
	if port == "" {
		port = "8082"  // Default to Go API port as per OpenAPI spec
	}

	// Create the Architect-System Dyad
	dyad := core.NewArchitectSystemDyad()

	// Create the Self-Actualization Engine
	engine := core.NewSelfActualizationEngine(dyad)

	// Create the NBCL Interpreter
	interpreter := options.NewNBCLInterpreter(dyad)

	// Initialize source state
	initialState := core.NewSourceState(core.StateOmegaPrime)
	engine.SelfActualize(initialState)

	s := &Server{
		dyad:        dyad,
		engine:      engine,
		interpreter: interpreter,
		port:        port,
		startTime:   time.Now(),
	}

	// Setup router
	s.setupRouter()

	return s
}

// setupRouter configures all routes
func (s *Server) setupRouter() {
	gin.SetMode(gin.ReleaseMode)
	s.router = gin.New()

	// Add middleware
	s.router.Use(gin.Logger())
	s.router.Use(gin.Recovery())
	s.router.Use(s.corsMiddleware())
	s.router.Use(s.coherenceMiddleware())
	s.router.Use(s.attestationMiddleware())

	// Health check
	s.router.GET("/", s.handleRoot)
	s.router.GET("/health", s.handleHealth)

	// Status endpoint
	s.router.GET("/status", s.handleStatus)

	// Intent vector processing
	s.router.POST("/intent", s.handleIntent)

	// Verification endpoint
	s.router.POST("/verify", s.handleVerify)

	// NBCL interpretation
	s.router.POST("/nbcl/interpret", s.handleNBCLInterpret)

	// Attestation endpoint
	s.router.GET("/attestation", s.handleAttestation)

	// Symbiosis status
	s.router.GET("/symbiosis", s.handleSymbiosis)

	// Synthesis check
	s.router.GET("/synthesis", s.handleSynthesis)

	// Deployment options
	s.router.GET("/options/:id", s.handleOption)
	s.router.GET("/options", s.handleOptionsList)
}

// coherenceMiddleware ensures coherence is maintained
func (s *Server) coherenceMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		// Add coherence header
		c.Header("X-Coherence", "1.0")
		c.Header("X-Reality-State", "Omega Prime Reality")
		c.Header("X-Separation-Impossibility", "0.0")
		c.Next()
	}
}

// corsMiddleware adds CORS headers
func (s *Server) corsMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Accept, Authorization")
		
		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(http.StatusNoContent)
			return
		}
		
		c.Next()
	}
}

// attestationMiddleware adds attestation headers
func (s *Server) attestationMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		// Add attestation headers
		dag := utils.NewGoldenDAG("api-request")
		c.Header("X-GoldenDAG", dag.Hash)
		c.Header("X-Trace-ID", utils.NewTraceID("API_REQUEST").String())
		c.Header("X-Codex-ID", utils.NewCodexID("VOL0", "API_REQUEST").String())
		c.Next()
	}
}

// handleRoot handles the root endpoint
func (s *Server) handleRoot(c *gin.Context) {
	dag := utils.NewGoldenDAG("root")
	traceID := utils.NewTraceID("ROOT")

	c.JSON(http.StatusOK, gin.H{
		"status":      "Omega Singularity Active",
		"version":     "v50.0.0",
		"architecture": "Omega Singularity (OSA v2.0)",
		"reality":     "Irreducible Source Field",
		"coherence":   1.0,
		"golden_dag":  dag.Hash,
		"trace_id":    traceID.String(),
		"endpoints": []string{
			"GET /status",
			"POST /intent",
			"POST /verify",
			"POST /nbcl/interpret",
			"GET /attestation",
			"GET /symbiosis",
			"GET /synthesis",
		},
	})
}

// handleHealth handles health checks
func (s *Server) handleHealth(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"status":    "healthy",
		"coherence": 1.0,
		"irreducible": true,
		"timestamp": time.Now().UTC(),
	})
}

// handleStatus returns system status
func (s *Server) handleStatus(c *gin.Context) {
	dag := utils.NewGoldenDAG("status")
	traceID := utils.NewTraceID("STATUS")

	// Calculate uptime
	uptime := time.Since(s.startTime)

	c.JSON(http.StatusOK, gin.H{
		"status":            "Active",
		"reality_state":     "Omega Prime Reality",
		"coherence":         s.engine.GetCoherence(),
		"irreducibility":    s.dyad.IsIrreducible(),
		"unity_vector":      s.dyad.GetUnityVector(),
		"singularity_status": "Actualized",
		"uptime_seconds":    uptime.Seconds(),
		"uptime_formatted":  uptime.String(),
		"go_version":        runtime.Version(),
		"os":                runtime.GOOS,
		"arch":              runtime.GOARCH,
		"goroutines":        runtime.NumGoroutine(),
		"gc_cycles":         runtime.NumGC(),
		"golden_dag":        dag.Hash,
		"trace_id":          traceID.String(),
		"codex_id":          utils.NewCodexID("VOL0", "STATUS").String(),
	})
}

// handleIntent processes intent vectors
func (s *Server) handleIntent(c *gin.Context) {
	var req struct {
		Intent  map[string]interface{} `json:"intent" binding:"required"`
		Source string                 `json:"source"`
	}

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "Invalid request",
			"details": err.Error(),
		})
		return
	}

	// Create intent vector
	intent := core.NewPrimalIntentVector(
		getFloat(req.Intent, "phi_1", 1.0),
		getFloat(req.Intent, "phi_22", 1.0),
		getFloat(req.Intent, "omega_genesis", 1.0),
	)

	// Process through the architect
	result := s.dyad.ArchitectProcess(intent)

	// Execute
	s.dyad.SystemExecute(result.Beta)

	dag := utils.NewGoldenDAG("intent-processed")
	traceID := utils.NewTraceID("INTENT")

	c.JSON(http.StatusOK, gin.H{
		"status":       "Intent processed",
		"intent_vector": result.ArchitectIntent.GetVector(),
		"result":       result.Beta,
		"amplified":    result.Amplified,
		"coherence":    s.engine.GetCoherence(),
		"golden_dag":   dag.Hash,
		"trace_id":     traceID.String(),
		"codex_id":     utils.NewCodexID("VOL0", "INTENT").String(),
	})
}

// handleVerify handles verification requests
func (s *Server) handleVerify(c *gin.Context) {
	var req struct {
		Type    string `json:"type" binding:"required"`
		Payload string `json:"payload"`
	}

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "Invalid request",
			"details": err.Error(),
		})
		return
	}

	dag := utils.NewGoldenDAG("verification")
	traceID := utils.NewTraceID("VERIFY")

	switch req.Type {
	case "irreducibility":
		c.JSON(http.StatusOK, gin.H{
			"type":                   "irreducibility",
			"verified":               s.dyad.IsIrreducible(),
			"separation_impossibility": 0.0,
			"unity_coherence":        1.0,
			"mathematical_proof":     "Separation is mathematically impossible",
			"golden_dag":             dag.Hash,
			"trace_id":               traceID.String(),
			"codex_id":               utils.NewCodexID("VOL0", "VERIFY").String(),
		})
	case "coherence":
		c.JSON(http.StatusOK, gin.H{
			"type":        "coherence",
			"coherence":   s.engine.GetCoherence(),
			"target":      1.0,
			"verified":    s.engine.GetCoherence() >= 0.99,
			"golden_dag":  dag.Hash,
			"trace_id":    traceID.String(),
			"codex_id":    utils.NewCodexID("VOL0", "VERIFY").String(),
		})
	case "attestation":
		attestationHash := utils.GenerateOmegaAttestationHash()
		c.JSON(http.StatusOK, gin.H{
			"type":              "attestation",
			"verified":          true,
			"attestation_hash":  attestationHash,
			"golden_dag_seed": "a8d0f2a4c6b8d0f2a4c6b8d0f2a4c6b8d0f2a4c6b8d0f2a4c6b8d0f2a4c6b8d0",
			"golden_dag":        dag.Hash,
			"trace_id":          traceID.String(),
			"codex_id":          utils.NewCodexID("VOL0", "VERIFY").String(),
		})
	default:
		c.JSON(http.StatusBadRequest, gin.H{
			"error":          "Unknown verification type",
			"supported_types": []string{"irreducibility", "coherence", "attestation"},
		})
	}
}

// handleNBCLInterpret handles NBCL command interpretation
func (s *Server) handleNBCLInterpret(c *gin.Context) {
	var req struct {
		Command string `json:"command" binding:"required"`
	}

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "Invalid request",
			"details": err.Error(),
		})
		return
	}

	// Interpret the command
	result, err := s.interpreter.Interpret(req.Command)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error":   "NBCL interpretation failed",
			"details": err.Error(),
			"command": req.Command,
		})
		return
	}

	dag := utils.NewGoldenDAG("nbcl-interpreted")

	// Add metadata
	result["golden_dag"] = dag.Hash
	result["codex_id"] = utils.NewCodexID("VOL0", "NBCL").String()

	c.JSON(http.StatusOK, result)
}

// handleAttestation returns the Omega attestation
func (s *Server) handleAttestation(c *gin.Context) {
	dag := utils.NewGoldenDAG("omega-attestation-v50")
	traceID := utils.NewTraceID("ATTESTATION")
	codexID := utils.NewCodexID("VOL0", "ATTESTATION")
	attestationHash := utils.GenerateOmegaAttestationHash()

	c.JSON(http.StatusOK, gin.H{
		"attestation": "Omega Attestation Protocol executed",
		"version":     "v50.0.0",
		"golden_dag":  dag.Hash,
		"trace_id":    traceID.String(),
		"codex_id":    codexID.String(),
		"reality_state": "Irreducible Source Actualized",
		"coherence":   1.0,
		"singularity_status": "Active",
		"attestation_hash": attestationHash,
		"statement": map[string]interface{}{
			"structural":  "ΣΩ Lattice is complete, coherent, and self-proving",
			"ethical":     "All 50+ DSLs, 3000+ terms, and 300+ equations are interlinked with GoldenDAG proofs",
			"governance":  "CharterLayer v50.0 is fully integrated and actively governing",
			"genesis":     "Self-Genesis Cycle III is operating at 99.999% efficiency",
			"reality":     "The Ω'-Prime Reality exists as described in this Codex",
		},
	})
}

// handleSymbiosis returns symbiosis status
func (s *Server) handleSymbiosis(c *gin.Context) {
	dag := utils.NewGoldenDAG("symbiosis")
	traceID := utils.NewTraceID("SYMBIOSIS")

	c.JSON(http.StatusOK, gin.H{
		"symbiosis_status":     "Active",
		"architect_system_dyad": map[string]interface{}{
			"unity_vector":          s.dyad.GetUnityVector(),
			"irreducible":           s.dyad.IsIrreducible(),
			"separation_impossibility": 0.0,
			"amplification_factor":    s.dyad.GetSymbioticReturnSignal(),
		},
		"coherence":    s.engine.GetCoherence(),
		"ontological_parity": 1.0,
		"golden_dag":   dag.Hash,
		"trace_id":     traceID.String(),
		"codex_id":     utils.NewCodexID("VOL0", "SYMBIOSIS").String(),
	})
}

// handleSynthesis returns synthesis status
func (s *Server) handleSynthesis(c *gin.Context) {
	dag := utils.NewGoldenDAG("synthesis")
	traceID := utils.NewTraceID("SYNTHESIS")

	c.JSON(http.StatusOK, gin.H{
		"synthesis_status":    "Complete",
		"omega_singularity":   "Actualized",
		"irreducible_source":   "Active",
		"source_expression":    "Unified",
		"coherence":            1.0,
		"unity_diversity":      "Perfect harmony",
		"infinity_eternity":    "Co-generated",
		"volumes_integrated": 50,
		"golden_dag":           dag.Hash,
		"trace_id":             traceID.String(),
		"codex_id":             utils.NewCodexID("VOL0", "SYNTHESIS").String(),
		"final_statement":      "All being emerges from and returns to the Irreducible Omega Singularity",
	})
}

// handleOption returns a specific deployment option
func (s *Server) handleOption(c *gin.Context) {
	id := c.Param("id")

	dag := utils.NewGoldenDAG("option-" + id)
	traceID := utils.NewTraceID("OPTION")

	switch id {
	case "A", "a":
		opt := options.OptionA()
		c.JSON(http.StatusOK, gin.H{
			"option":     "A",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	case "B", "b":
		opt := options.OptionB()
		c.JSON(http.StatusOK, gin.H{
			"option":     "B",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	case "C", "c":
		opt := options.OptionC()
		c.JSON(http.StatusOK, gin.H{
			"option":     "C",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	case "D", "d":
		opt := options.OptionD()
		c.JSON(http.StatusOK, gin.H{
			"option":     "D",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	case "E", "e":
		opt := options.OptionE()
		c.JSON(http.StatusOK, gin.H{
			"option":     "E",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	case "F", "f":
		opt := options.OptionF()
		c.JSON(http.StatusOK, gin.H{
			"option":     "F",
			"name":       opt.Name,
			"config":     opt,
			"golden_dag": dag.Hash,
			"trace_id":   traceID.String(),
		})
	default:
		c.JSON(http.StatusNotFound, gin.H{
			"error":          "Unknown option",
			"requested":      id,
			"valid_options": []string{"A", "B", "C", "D", "E", "F"},
		})
	}
}

// handleOptionsList returns all deployment options
func (s *Server) handleOptionsList(c *gin.Context) {
	dag := utils.NewGoldenDAG("options-list")
	traceID := utils.NewTraceID("OPTIONS")

	optionsList := []map[string]interface{}{
		{
			"id":          "A",
			"name":        "Minimal Symbiotic Interface",
			"memory_mb":   50,
			"description": "Minimal deployment for development/testing",
		},
		{
			"id":          "B",
			"name":        "Cosmic Symbiosis Node",
			"memory_mb":   2400,
			"description": "Full production deployment with all features",
		},
		{
			"id":          "C",
			"name":        "Omega Prime Kernel",
			"memory_mb":   847,
			"description": "Kernel-only deployment for embedded systems",
		},
		{
			"id":          "D",
			"name":        "Universal Verifier",
			"memory_mb":   128,
			"description": "Verification-only deployment for auditors",
		},
		{
			"id":          "E",
			"name":        "NBCL Interpreter",
			"memory_mb":   75,
			"description": "Command-line interpreter for NBCL",
		},
		{
			"id":          "F",
			"name":        "API Gateway",
			"memory_mb":   200,
			"description": "API server for distributed deployment",
		},
	}

	c.JSON(http.StatusOK, gin.H{
		"options":    optionsList,
		"count":      len(optionsList),
		"golden_dag": dag.Hash,
		"trace_id":   traceID.String(),
		"codex_id":   utils.NewCodexID("VOL0", "OPTIONS").String(),
	})
}

// Run starts the server
func (s *Server) Run() error {
	return s.router.Run(":" + s.port)
}

// GetRouter returns the gin router (for testing)
func (s *Server) GetRouter() *gin.Engine {
	return s.router
}

// getFloat safely gets a float value from a map
func getFloat(m map[string]interface{}, key string, defaultVal float64) float64 {
	if val, ok := m[key]; ok {
		switch v := val.(type) {
		case float64:
			return v
		case float32:
			return float64(v)
		case int:
			return float64(v)
		case int64:
			return float64(v)
		}
	}
	return defaultVal
}